Create StringFormatter.cpp
While taking a break from studying for an hour, I decided to challenge myself with creating a string formatter. I think I could make it more simple, somehow using a secondary string instead of a char array but the char array seemed to make the most sense to me at the time.